pyproject.toml
[project]
name = "banking-langgraph-agent"
version = "0.1.0"
description = "LangGraph banking multi-agent backend with account, payments, and service support agents"
authors = [
    { name = "Banking Agent Team" },
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.124.0",
    "httpx>=0.28.1",
    "langchain>=1.1.2",
    "langchain-openai>=1.1.0",
    "langfuse>=3.10.5",
    "langgraph>=1.0.4",
    "langgraph-checkpoint-postgres>=3.0.1",
    "langgraph-swarm>=0.1.0",
    "psycopg-pool>=3.3.0",
    "psycopg2-binary>=2.9.11",
    "psycopg[binary]>=3.3.2",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "typing-extensions>=4.12.0",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.8",
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
